move decoder back to its own package
